:fire: Change footer & header file positions
Change footer and header files to /src/components, fix GlobalStyle.css that was affecting layout, and change <div> to <> on Home page
