+++ frontend/pages/index.js
@@ -1,75 +1,71 @@
 import React from 'react';
 import { Container, Row, Card, Button } from 'react-bootstrap';
 import styles from '../src/theme/pages/index.module.css';
-import Header from './components/header.js';
-import Footer from './components/footer.js';
+import Header from '../src/components/Header';
+import Footer from '../src/components/Footer';
 
 export default function Home() {
   return (
-    <div>
-      <Header/>
-      
+    <>
+      <Header />
       <Container>
-        <Container>
-          <Row className="justify-content-md-between">
-            <Card className={styles.smlCard}>
-              <Card.Body>
-                <Card.Title>Documentation</Card.Title>
-                <Card.Text>
-                  Find in-depth information about Next.js features and API.
-                </Card.Text>
-                <Button variant="primary" href="https://nextjs.org/docs">
-                  More &rarr;
-                </Button>
-              </Card.Body>
-            </Card>
-            <Card className={styles.smlCard}>
-              <Card.Body>
-                <Card.Title>Learn</Card.Title>
-                <Card.Text>
-                  Learn about Next.js in an interactive course with quizzes!
-                </Card.Text>
-                <Button variant="primary" href="https://nextjs.org/learn">
-                  More &rarr;
-                </Button>
-              </Card.Body>
-            </Card>
-          </Row>
-          <Row className="justify-content-md-between">
-            <Card className={styles.smlCard}>
-              <Card.Body>
-                <Card.Title>Examples</Card.Title>
-                <Card.Text>
-                  Discover and deploy boilerplate example Next.js projects.
-                </Card.Text>
-                <Button
-                  variant="primary"
-                  href="https://github.com/vercel/next.js/tree/master/examples"
-                >
-                  More &rarr;
-                </Button>
-              </Card.Body>
-            </Card>
-            <Card className={styles.smlCard}>
-              <Card.Body>
-                <Card.Title>Deploy</Card.Title>
-                <Card.Text>
-                  Instantly deploy your Next.js site to a public URL with
-                  Vercel.
-                </Card.Text>
-                <Button
-                  variant="primary"
-                  href="https://vercel.com/new?utm_source=github&utm_medium=example&utm_campaign=next-example"
-                >
-                  More &rarr;
-                </Button>
-              </Card.Body>
-            </Card>
-          </Row>
-        </Container>
+        <Row className="justify-content-md-between">
+          <Card className={styles.smlCard}>
+            <Card.Body>
+              <Card.Title>Documentation</Card.Title>
+              <Card.Text>
+                Find in-depth information about Next.js features and API.
+              </Card.Text>
+              <Button variant="primary" href="https://nextjs.org/docs">
+                More &rarr;
+              </Button>
+            </Card.Body>
+          </Card>
+          <Card className={styles.smlCard}>
+            <Card.Body>
+              <Card.Title>Learn</Card.Title>
+              <Card.Text>
+                Learn about Next.js in an interactive course with quizzes!
+              </Card.Text>
+              <Button variant="primary" href="https://nextjs.org/learn">
+                More &rarr;
+              </Button>
+            </Card.Body>
+          </Card>
+        </Row>
+        <Row className="justify-content-md-between">
+          <Card className={styles.smlCard}>
+            <Card.Body>
+              <Card.Title>Examples</Card.Title>
+              <Card.Text>
+                Discover and deploy boilerplate example Next.js projects.
+              </Card.Text>
+              <Button
+                variant="primary"
+                href="https://github.com/vercel/next.js/tree/master/examples"
+              >
+                More &rarr;
+              </Button>
+            </Card.Body>
+          </Card>
+          <Card className={styles.smlCard}>
+            <Card.Body>
+              <Card.Title>Deploy</Card.Title>
+              <Card.Text>
+                Instantly deploy your Next.js site to a public URL with Vercel.
+              </Card.Text>
+              <Button
+                variant="primary"
+                href="https://vercel.com/new?utm_source=github&utm_medium=example&utm_campaign=next-example"
+              >
+                More &rarr;
+              </Button>
+            </Card.Body>
+          </Card>
+        </Row>
       </Container>
-    
-      <Footer/>
-    </div>
+
+      <Footer />
+    </>
   );
 }
