+++ frontend/src/theme/GlobalStyle.css
@@ -13,5 +13,6 @@
 }
 
 body {
-  margin: 13rem 0rem 0rem 0rem;
+  margin: 0;
+  padding: 0;
 }
